refactor url validator
