+++ lib/record_validators/command_requests/image_to_video/picture_validator.rb
@@ -1,0 +1,19 @@
+module RecordValidators
+  module CommandRequests
+    class ImageToVideo
+      class PictureValidator
+        def initialize(picture_id:)
+          @picture_id = picture_id
+        end
+
+        def valid?
+          picture_id.present?
+        end
+
+        private
+
+        attr_reader :picture_id
+      end
+    end
+  end
+end
